updated pricing section
