+++ app/page.tsx
@@ -2,7 +2,7 @@
 import Accordian from "@/components/Accordian";
 import ContactPage from "@/components/Contact";
 import HeroSection from "@/components/main/Hero";
-import PricingSection from "@/components/PricingSection";
+import PricingSection from "@/components/Pricing/PricingSection";
 
 import Services from "@/components/Services";
 import TeamPage from "@/components/Team";
